# Tests and performance fixes for hash helpers.