--- a/PwnedPasswords.sln
+++ b/PwnedPasswords.sln
@@ -10,6 +10,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{E7E72F74-A4A4-4D3F-8F97-703FA172647C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -20,6 +22,10 @@ Global
 		{ED82AA3F-3700-4C81-8393-89270437C908}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ED82AA3F-3700-4C81-8393-89270437C908}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ED82AA3F-3700-4C81-8393-89270437C908}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E7E72F74-A4A4-4D3F-8F97-703FA172647C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7E72F74-A4A4-4D3F-8F97-703FA172647C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E7E72F74-A4A4-4D3F-8F97-703FA172647C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E7E72F74-A4A4-4D3F-8F97-703FA172647C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE