@mdawsonuk 2021-05-31
@stebet Awesome work! I was wondering who would add tests first 😁 Hopefully this is merged soon so I can update the NTLM hash function on my PR